important progress for 0.2
 * introduce Ld_Hasher for hashing passwords (based on phpass)
 * introduce cookie and cookie session storage classes
 * make authStorage configurable for sessions
 * start to replace constants (LD_CSS_DIR, etc) by site properties
 * introduce $instance->getDbConnection()
 * file_put_contents() -> Ld_Files::put()
 * support 'user' preference type
 * move users functions from Ld_Auth to Ld_Site


git-svn-id: http://ladistribution.net/svn/trunk@42 ae0f9bec-b264-4796-a651-e028ba2a50ff

+++ library/Ld/Repository/Local.php
@@ -139,7 +139,7 @@ class Ld_Repository_Local extends Ld_Repository_Abstract
     {
     }
 
-    public function importPackage($filename)
+    public function importPackage($filename, $clean = true)
     {
         $package = new Ld_Package(array('zip' => $filename));
 
@@ -151,9 +151,13 @@ class Ld_Repository_Local extends Ld_Repository_Abstract
 
         Ld_Files::put($dir . "/manifest.xml", $package->getManifestXml());
 
-        Ld_Files::unlink($filename);
+        if ($clean) {
+            Ld_Files::unlink($filename);
+        }
 
         $this->generatePackageList();
+
+        return $package;
     }
 
     public function getDirectory($params)
